moved news component to home
