+++ packages/client/src/app/pages/home/index.ts
@@ -1,2 +1,3 @@
 export { HomeComponent } from './home.component';
+export { NewsComponent } from './news/news.component';
 export { WatchingComponent as WatchingPageComponent } from './watching/watching.component';
